morebits: Fix bug defaulting followCrossNsRedirect to false
Introduced in #915, but if the second parameter to `setFollowRedirect` (aka `followCrossNsRedirect`) isn't provided, it is passed as `undefined` aka falsey.  That breaks anything that is was using `setFollowRedirect` without explicitly passing a second parameter (aka everything), given that the default value is `true`.

As elsewhere, this `typeof` style is because IE11 doesn't support default parameters.