warn: Fix bug in autolevel not elevating level for singlewarn
Was using the default selected template rather than the latest warning found
in the page object, meaning the "found" level would be set to 1, so even if a
singlewarn template was found, it wouldn't be elevated.  If you selected a
singlewarn template, then went to the autolevel menu, the default would *be* a
singlewarn, so it would appear to be processed properly.  Fixes #1027.