block: Add DS topic codes for aeblock/aepblock
The restructuring at the end of `change_action` has means `change_preset` is additionally being called whenever blocking is happening, not just if in the context of also applying a template.  This leads to `change_template`, which toggles the visibility of the DS select.

This has the side effect of changing how the form behaves when a user is already blocked.  Right now, the form doesn't display any of the old options by default (block and issue template).  *If* the sysop unchecks the issue template option, the block form will prefill with all the information from the original block (expiry, reason, etc.).  It's ephemeral, though, as if a preset is chosen, that information is then wiped clean, only returning if the issue template checkbox is toggled.

That's pretty misleading/confusing, especially since no one should be expected to memorize the information.  The change to `change_action` means that, when the template box is unchecked, the old information is not prefilled, matching the behavior elsewhere.  If we want to *always* prefill, that's a separate discussion.