block: Favor current block information if present
Before 91ccacf, the block form would inconsistently and rarely load the prior options; I know folks were occasionally confused as to what was happening.  That commit made it never do so, as a measure of consistency.  This commit adds the prior block conditions as a new preset, which is the default selection upon form loading or action toggling.  This should restore the behavior but make it more explicit when the information is present, and more importantly allow the sysop to reliably *go back* to the prior block settins.

I've also toggled the initial loading of the form to account for partial status, although the preset itself is dynamically adjusted to account for `{{uw-block}}`, `{{uw-pblockindex}}`, etc. as appropriate.  Additionally, I've explicitly added `bkprops` so we can preload any partial block restrictions, as those were missing.