protect: Ensure a valid type of move protection
Since move tries to match edit if present, but autoconfirmed is not an option (see discussion in #451), it's possible for someone to attempt a submission without any level (sysop, template, etc.) selected for move protection if they select move protection while the default edit level of "autoconfirmed" is selected, and then uncheck edit protection.  The query would just fail, but better not to reset the form.

Also moved the empty default selection for filing a request to the top, as might be expected.