(batch)protect: consolidate near-identical expiry objects into Twinkle.protect.protectionLengths
`changePreset` was being used for edit and move, so we can just use it for PC and create protection as well.  The menus now always revert to their default value when selected, where previously only items that were *not* automatically enabled (e.g. edit protection when changing to a move preset) would retain their previous value.

These were tweaked around in 9c29d52c and 7b09cb46 back in 2016, but the net change is move was made to default to 2 days, not indefinite; that's retained here.