fluff: Revert multiple hidden revisions in a row
Assume that two revisions in a row from a revdel'd user are the same user, as it's most likely the case.  This also makes things smarter for sysops, as even if they don't see the name when clicking the link, the query will fetch it for them.

Also defaults `vandal` (the username) in `rollbacklinks` to `null` to match a `rev[i].getAttribute('user')` on a hidden name.