xfd: Begin basic autoedit request function
Currently written with repetitive `if (pageobj.canEdit())` loop.  That's just because everyone but AfD has a distinct function on page load, but AfD is muddled within a bunch of other checks.  Those checks are good (and should maybe be added to the others...), so we would have to break out the actual tagging into a separate function that doesn't need more loading to cut down on the ifs while still handling the loading for everyone else.  PITA.

Would close #176