xfd: Don't show reason for AfD errors that don't matter
Both delsort and pagetriage can lead to errors; the former mainly for nonexistent categories (rare), the latter for pages not in the triage queue (common, see #930).  None of these matter for the sake of losing the nomination reason, so neither needs to get evaluated before `Twinkle.xfd.currentRationale` is `null`ed.  Likely to be fairly common for the `triage` action.

If we handled `triage` the way initially proposed in #930, we wouldn't see these errors, so this is kind of kludgy.  Probably still worth it.