xfd: Use a unified user notification function
With all the XfD notices unified (29ae3373), we don't need to handle them individually.  This handles all current notifications, including without double-logging on MfDs, TfMs, and RfD, and handles non-userspace targets in RfD target notifications; the majority of the complexity is due to handling MfD and RfD.  Should also work for #1001.

Also fixes a bug in MfD notifications, where you could only notify the userspace owner if you were also notifying the creator, regardless of whether they were the same or not.