morebits: fix error handling on no internet
errorCode is `null` on no internet connectivity, not `undefined` let alone the string `'undefined'` which the code was looking for.
This bug was also preventing any onFailure handler from triggering because of the TypeError on line 3260 (null doesn't have `indexOf` method). Converted hard error handling code to use a switch.