fluff: Fix bug in identifying revid
This has been comparing a revid against a revision's `pageId` for over a decade, but `pageId` isn't part of a revision (or anything) and `pageid` would be the wrong thing to compare anyway!  I *think* it's possible there was a `pageId` parameter being returned back in the day, but there certainly isn't now.