morebits: Don't overwrite unprovided edit protection with cascading
If someone tried to set just move protection on an unprotected page that was covered by cascading protection, we were picking up that cascading protection and pushing it into the query.  This could lead to some weird results - oddly mismatched edit/move protection, the highest protection level when none was intended - but of course it has literally never mattered since the page was covered by cascading protection anyway.  It *could* theoretically matter if cascading protection was removed down the line, but at any rate, we might as well be correct.