morebits: Enforce newer API error message format, defaulting to html
Defaults to `html`, but `plaintext` and `wikitext` are acceptable; `raw`, `none`, and the legacy format `bc` will be converted to `html`.  Currently limited to the first error found, which was the case with the old syntax and is assumed by our uses of `errorCode`/`errorText`.

Of note, this also defaults all queries to the wiki's `content` language, see discussion in #1179.

https://www.mediawiki.org/wiki/API:Errors_and_warnings