tag: reorganize: use "parameter" in subgroup to make template wikitext
Instead of a big switch to fill out template parameters, the name of the parameter is stored in the subgroup and then we loop thorough the subgroups of the given tag to create the wikitext.

The few instances where we used to add things like "|1=article" are handled by using a hidden Morebits input with parameter=1 and value=article.