block: Don't add expiry to template if it's a date, not a duration
The block templates ask for a duration in the `time` parameter, not a date: someone manually set an actual date as the block expiry, which is acceptable to MediaWiki, the block template would say something like `...blocked from editing for a period of April 1, 1984 13:00`.  This simply excludes the duration if the string is a valid date (thus not a duration).