morebits: date: cleanup constructor (#1187)
Avoid putting in special cases to account for myriad browser differences – that isn't resilient or future-proof. The constructor was only meant to parse the signature timestamp format in addition to whatever Date() parses. Making this explicit by extracting the date parts from the string and using `Date.UTC(year, month, date, hour, minute)` which returns the milliseconds since epoch, which is then used to make the Date object. Both of these are part of the standard[1] and thus guaranteed to work everywhere.

[1]: https://www.ecma-international.org/ecma-262/5.1/#sec-15.9.4.3
and https://www.ecma-international.org/ecma-262/5.1/#sec-15.9.3.2