block/protect: Show expiry/timestamp for last action (if not currently restricted)
Shows timestamp for an unblock or unprotection action, and expiry for a block or protection, but only if the user/page is not currently blocked/protected.  For protection it's not particularly smart about finding the last expiry, it just grabs the first level, which defaults to edit.  That means that if the last action was to adjust the edit protection to expire before move protection expired, this would show the wrong one.