Detect RfD nomination in Morebits.wiki.isPageRedirect, use in CSD
It's been long desired to show the redirect CSDs on redirects nominated at RfD.  This adds detection of that box to `Morebits.wiki.isPageRedirect` and uses it in twinklespeedy.  Should also improve things elsewhere that it's used, in particular redirect tagging in the tag module.  XfD will have problems with it, namely finding the redirect target for RfD.