morebits.date: Fix constructor order for firefox
Firefox handles the enWiki timestamp format but incorrectly (#936), in particular by accepting the comma, so the date is treated as local since firefox (and chrome) ignore the trailing `(UTC)`.  Somehow in testing #1187, we missed the fact that the native date will preempty the timestamp check, so Firefox is once again messing things up.