xfd: Don't try to add edit request if page is already a talk
Also add a note if the edit request fails.  These are unlikely but possible.  `Twinkle.xfd.currentRationale` will still be printed, but we don't want to null it out since there's no guarantee other important actions (namely, listing) have taken place, so we're stuck between a rock and a hard place until that whole thing is dealt with.