fluff: Only try to show oldid links on pages that actually exist
Viewing an oldid of a deleted page would produce an error, since it was likely editable and `wgCurRevisionId` and `wgArticleId` are both defined on `title=&oldid=844151270` urls, whereas `wgRevisionId` isn't.  It was also possible to trigger this when reviewing pending changes, although I haven't been able to replicate that for months.