tag: fix minor bug (from #1214)
Affects {{Expand language}} only. This tag is weird in that for some reason we want to add the parameter 1= with no value into the wikitext. #1214 has a somewhat hacky way to accommodate this situation. But it had a bug in that undefined was being put into the wikitext instead of the empty string. That is, it was adding `{{Expand language|1=undefined}}` when `{{Expand language|1=}}` is desired.