tag: fix minor bug (from #1216) (#1227)
Affects {{Expand language}} only. This tag is weird in that for some reason we want to add the parameter topic= with no value into the wikitext. #1216 has a somewhat hacky way to accommodate this situation. But it had a bug in that undefined was being put into the wikitext instead of the empty string. That is, it was adding `{{Expand language|topic=undefined}}` when `{{Expand language|topic=}}` is desired.