morebits.wiki.page: Get watched status of page, only apply watchlist expiry if unwatched
Watchlist expiry is ideal for Twinkle since a lot of edits or taggings are transitory.  If, however, you have a page you've watched indefinitely, the API will overwrite that with the expiry settings, which is generally *not* what folks want.  Until https://phabricator.wikimedia.org/T268834, there's no reliable or convenient way to query an individual page's watchlist expiration, so for now we rely simply on whether the page is being watched or not.  It's not ideal, but especially early on, is probably a fair assumption.  Exposed via `getWatched`.

I had thought I tested this before rollout, but it was being masked by a silly bug: for `action=watch`, it's `expiry` not `watchlistexpiry`.  Regardless, that's now fixed for the ad-hoc watching mechanisms in xfd (`mb.w.page`) and fluff (`mb.w.api`), and the above technique used as well.