mb.wikitext.page: Fix ancient bug where multiple occurences were not being processed
`string.replace(text, '', 'g')` has been present in `removeTemplate`, `commentOutImage`, and `addToImageComment` for ages, but it's not a valid form of `string.replace`.  The `'g'` was being entirely ignored!  Instead of creating a `new RegExp` relying on the output from `splitWeightedByKeys`, we can just not remove the duplicated items; thus, each item refers to a single element needing action.

In theory I'd be a little wary about changing these given the off chance something was relying on them not working fully - although the initial capitalization changes would have been processed - but it's clear the intent of these was to work, so I think it's fine.  Besides, there aren't many non-Twinkle uses.  `removeTemplate` probably worked just fine if the "template" was in another namespaces, as this generally worked fine for `commentOutImage` and `addToImageComment` since they included the namespace.