morebits.pageNameRegex: Properly escape entire string, use in Morebits.wikitext.page
There's more to matching a page name than just the first character; this applies `Morebits.string.escapeRegExp` (spaces/underscores to `[_ ]` and all the loveliness from `mw.util.escapeRegExp`) as well.  Underutilized since its inception, it's only used in the tag module, so I don't think there should be anything relying on it *not* matching pages it should be.  It is, however, perfect for the `Morebits.wikitext.page` methods.

Casing is largely handled with to `mw.Title.phpCharToUpper`, which necessitated introducing the whole of phpCharToUpper.json and is our first mocking of `mw.Title`.