protect: Don't overwrite indefinite watching with expiry
PR #1133 was merged without fully accounting for the fact that providing a watchlist expiry would overwrite indefinite watching, which is nonideal; see #1230.  Like #1230, this is imperfect, as it only provides a boolean for watched or not, and doesn't give any information on an expiry.  We can at least rely on the information being available, though.