morebits: Force loading of a page if a watchlist expiry is provided
Tried a few different things, but after thinking about it for a bit longer, I think this is just the only really viable option, at least unless/until the information is exposed in an easier way (T270057).  Would largely revert #1241 as unnecessary, opting to build on/expand #1230.  Still imperfect since there's no information about whether the page is watched indefinitely or with an expiry (T268834), but at least this means we'll never overwrite indefinite with expiry; it also means we won't extend an expiration.

To do *that*, right now I think the only way would be to rewrite everything to not use the `watchlist` option, but rather a new callback with `action=watch`, although that doesn't account for userpreferences well, and would, of course, be awful.