Morebits.namespaceRegex for matching all namespace capitalizations and aliases, use in morebits and modules
We were manually checking namespaces, which was largely working fine, but was repetitive and prone to ignoring case (`Morebits.wikitext.page` methods) or alternative aliases (`Project`).  A string is more versatile than an actual regex string/RegExp object, but means that it gets fairly awkward; hence this function.  Used in file-specific regex in `Morebits.wikitext.page` methods, for which it enables `commentOutImage` and `addToImageComment` to parse images regardless of namespace or title capitalization, and in speedy, xfd, image, batchdelete, talkback, where there should be no negative change.

In theory, this could include an option to toggle the trailing colon - even defaulting to off, since everywhere we'd use this currently uses it - but it could some unwanted complications around mainspace.  That is, `:` isn't a desirable regex for mainspace, but defaulting to including the leading colon isn't always desirable (categories, files), and something like `0,1` giving `(?::?|(?:[Tt]alk):)` is awkward.  Then again, it's probably rare that a regex is being used like this for mainspace, so maybe it'd be fine?