tests: Account for a few missing cases in removeLink
Include category links as well as links with display text versus those without.

Also moved an erroneous comment out of jsdocs back to where it belongs.  Moved in a462aeec but it was in reference to the colon-switching, not the function itself.  No changes occurred to the function, but it was misleading.