block: Enable for IP ranges, improve handling of rangeblocks
- Enable on ranges
    - Leaving a template is disabled, but the "templated reason" {{rangeblock}} is used as the default
- Determine if IP is caught in a range block
    - Display the range, whether it's the same block, etc.
- Use correct information for IPs covered by two blocks (e.g. a single IP and a range block)
- Move/hide unblock link for IPs covered by ranges but not directly blocked

A lot of this largely hinges on "is this user the same as the user showing up in the blocks API," with the main trick being the use of `bkip` for IPs (range or otherwise), reserving `bkusers` for registered folks.

There's a fundamental MW issue when processing previous blocks (see [[phab:T270737]] and [[phab:T146628]]) where logs don't account for a rangeblock if it's not exact.  Briefly, blocks handles functionally-equivalent IP ranges nicely, but logs don't, so we can't ever be sure whether two ranges are the same.  IPUtils has the functionality, but it's a *ton* of extra junk for something minor that we can ameliorate that should instead be fixed upstream.

Part of #775