block: Add checkbox to just block the /64
Obviously relies on `Morebits.get64`, but otherwise, a large part of this is just making increased use of `relevantUserName` (rather than `mw.config.get('wgRelevantUserName')`), and simply adjusting that when (un)checking the checkbox.  A number of things previously stable need to be changeable, but the overall structure isn't too different, since this form is used to reloading frequently.  The main things adjusted:

- Unblock footerlink and window title toggled
- Template option unchecked/disabled as necessary
- Prior block preset reset
- Data processing in `Twinkle.block.fetchUserInfo` is pulled out into `Twinkle.block.processUserInfo`
    - `Twinkle.block.processUserInfo` caches its results in `Twinkle.block.fetchedData`, thus allowing `Twinkle.block.fetchUserInfo` to only be rerun as necessary on multiple togglings

Closes #1173.