speedy/image: don't share addToLog
Speedy module needs to be changed a lot for localizability. It's no longer wise for image module to piggyback on speedy's `addToLog` function. This means some duplication but not much (particularly since #878). The integration was never elegant to begin with (noted in a comment as "params data structure when coming from DI is quite different"), so there was a large `if` separating the behaviour for the two modules.