Create promisified wrapper around Morebits.wiki.page
While very hacky, this will be immensely useful. The key advantage of using promises over deferreds is this:

```
let def = $.Deferred();
pageobj.load(() => {
  // ... Some TypeError here
  def.resolve();
});
return def;
```

If the TypeError occurs, the deferred object never gets resolved or rejected, and the function actually returns `undefined` (rather than a deferred), causing issues in error handling.

On the other hand, while using promises,
```
return pageobj.load(() => {
  // ... TypeError here
});
```

A promise is still returned, rejected with the TypeError.

This presets the "Twinkle" change tag, so that it doesn't have to be repeated on every action.