Morebits.validCIDR: test if IP range within CIDR limits; use to limit block and arv
With [T206954](https://phabricator.wikimedia.org/T206954), IP ranges that are too large will still show `wgRelevantUserName` as if anything can be utilized on them.  This adds `Morebits.validCIDR`, which returns false unless given a valid range within the CIDR limits.  Limits are hardcoded (/16 for IPv4 and /32 for IPv6) since `$wgBlockCIDRLimit` isn't exposed AFAICT.

It is then used to limit loading of arv and block for IP ranges.

With this adding to the new `get64` and the old `sanitizeIPv6`, I wonder if it'd be worth reorganizing under `Morebits.IPUtils` or something?