morebits: fnApplyWatchlistExpiry to decide whether/when to apply a watchlistexpiry
Basically, only attempt a watchlist expiry if (1) the page isn't watched, or (2) the page is already watched temporarily; in the latter case, try to only extend an existing temporary watching.

Tries not to be too over-enthusiastic, as there are plenty of valid MW expiry strings that `Morebits.date` doesn't recognize or can't `add` with.