protect: Fetch watched status of the page to properly watch when requesting
I was using the `Morebits.wiki.page` object returned from the RfPP page itself to determine watchlist expiry status of the tagged page, which is clearly wrong.  Since we're not loading the page directly in that case, we can just tack on to the protection fetching and store it for later.  Fixes #1315.

These whole `watch_query` things are annoying.