morebits: Expand setWatchlist handling of expiry, default to infinity for watch/preferences
`setWatchlist` now takes a second parameter, allowing you to explicitly set the watchlist expiry if so desired.

To break it down, the MediaWiki API allows four watching options:

- `watch`: Explicitly watch the page, using the expiry if provided
- `preferences`: Watch (or not) the page according to preferences, using the expiry if appropriate and provided
- `nochange` (default): Don't change the binary watch status, but if an expiry is provided and appropriate, use it
- `unwatch`: Expiry doesn't matter

We've made two small editorial decisions here:
- If no expiry is provided for a `watch` or `preferences` option, default to indefinite watching.  This mirrors prior (and expected) behavior.
- If `nochange` is provided, explicitly set the expiry to `null`, so `nochange` really means "no change."

The former is a bugfix for us: if a user wanted to indefinitely watch a temporarily watched page, we weren't handling that possibility just by using `setWatchlist` in Twinkle.  For the latter, `setWatchlistExpiry` continues to exist and remains useful.  This all - including `setWatchlist` falling back to treating a string as `watch` + expiry - is a bit kludgy in order to make Twinkle prefs work smoothly, but continues to make things easier.  Passing a falsey value for `watchlistExpiry` will get the expiry treated as nonexistent (with `fnApplyWatchlistExpiry` from #1301), using whatever exists on-wiki.