fluff: Dogfood fluff to use Morebits.wiki.page's revert editmode
With `Morebits.wiki.page.revert` fixed/improved, we can (finally) use it here to simplify things.  `revertToRevision` needs a `load` to snag the user for `Twinkle.fluff.formatSummary`, but otherwise the whole process is straightforward what `Morebits.wiki.page.revert` was designed for.  The rollback links of course still need the direct API query in order to get and process the slew of revisions, but we can simplify processing quite a bit, and finally remove the duplication in the completion callback.