morebits: render wikilink syntax in status and quickForm labels
Makes it much easier to include wikilinks in status messages and quickForm labels. Also de-duplicates some code in the two clases.

If literal `[[...]]` are to be included in status for whatever reason, that can still be done using the HTML entities `&#91;&#91;...&#93;&#93;`

$.parseHTML() removes <script> tags in the parsed output, so there should be no security issues.

A `Morebits.createHtml` function is added. This obsoletes `Morebits.status.prototype.codify` which is hence removed.