add i18n support for morebits (#1354)
Adds `Morebits.i18n` which allows plugging in an i18n library through which localised messages can be provided.

The English fallbacks continue to be part of the source code so for using morebits in English no i18n library is required. That is, the changes here are backward-compatible.

Messages in code now become:
```js
msg('msg-name', ...messageParams, fallbackEnglishMessage)
```

where the `fallbackEnglishMessage` is a string that directly uses the message params. That is, it is
```js
msg('deleting-pages', numPages, 'Deleting ' + numPages + ' page' + (numPages > 1 ? 's ' : ' '))
```
instead of
```js
msg('deleting-pages', numPages, 'Deleting $1 {{plural:$1|page|pages}}'
```
The latter would have been nicer but for that I'd have to rewrite all the messages and also then an i18n library would be required to use morebits. Maybe something to think about for the future – but not required for the twinkle-core project.

In case of some messages where a MediaWiki message exists - such as the month names - our message names match the ones in mediawiki.

I've skipped the messages that only the programmers should see -- such as parameter validations for functions, isSysop check for admin-only operations, etc.