morebits: add $data attribute for quickForm elements
Till now there has been the the quickForm extra attribute which allowed attaching arbitrary data to DOM nodes. But this is discouraged for many reasons (https://stackoverflow.com/questions/7895560/javascript-dom-setting-custom-dom-element-properties, https://stackoverflow.com/questions/3095336/is-it-bad-practice-to-add-properties-to-dom-nodes).

Using JQuery data (https://api.jquery.com/data/) is better.

Support for `extra` attribute is nevertheless retained, though this is unused in Twinkle.