morebits: put 3px margin after input labels
For almost every input and select field label, such a margin was being artificially added till now by ending the label with a trailing space.

However, trailing spaces are not allowed on translatewiki.net (technical limitation) and so needs to be written as `&#32;` (https://phabricator.wikimedia.org/T278627#7019324) which unnecessarily obfuscates so many messages.