Add support for range talk
Add support for pages that are a redirect (which is often done when IPv6-ranges are into play). The code will now follow that redirect, and drop the block message there. Following redirects to other namespaces is disabled for security reasons and because a talk page should not be in another space.