tag: better detection of "not marked as reviewed" (#1667)
* tag: better detection of "not marked as reviewed"

Fix #1573

Not working yet. Because of async, form is rendered before the "Mark the page as patrolled/reviewed" checkbox is appended.

Perhaps the API query needs to be set up to run before friendlytagCallback(). Any suggestions on how to set that up? I'm good at async/await, but not so good at promises and callbacks.

* fix rendering issue