Fix jest error message on localhost
Before this patch, when I tried to run `npm test`, I was getting the error "Test environment jest-environment-jsdom cannot be found. Make sure the testEnvironment configuration option points to an existing node module."

Stack Overflow and the error message suggest that installing the node package "jest-environment-jsdom" is the fix.

After installing this, the tests run, but a bunch of them fail on localhost. I plan to address that in a follow up patch.