speedy: refactor data to be in JSON
related #2199 refactor CSD code to make it easy to add CSDs

First of several patches. End goal is to make it so that to add/edit/remove a CSD, you only have to edit one spot in the code, instead of a bunch of different spots. This spot is in JSON format.

* add snapshot tests for everything we plan on refactoring
* fix twinklespeedy snapshot test name to be twinklespeedy instead of twinkleblock (copy paste error)
* refactor twinkleconfig to auto calculate preference order. this results in 3 less spots we need to manually update when adding a CSD
* update DEVELOPER.md to reflect the simpler process to add a CSD
* remove some excessive detail in DEVELOPER.md about creating new CSD lists (this is pretty rare)
* add a section to DEVELOPER.md about Twinkle's load order and how that works