warn: add {{uw-orphantalk}} (#2207)
This adds a warning for following-up on a G8 deletion of a mainspace
talk page created that had no article. It's intended for the
circumstance where there wasn't enough content to justify for example
moving to draftspace, but you want to let the editor know why it was
marked for deletion.

This is intentionally not done as part of notifying a user of a CSD
because 1) from what I can tell there's some hardcoding there to use the
much more aggressive db-reason-notice styled templates and 2) because
personal preference in how I've used this warning template since I've
added it to my config.