eslint: switch to camelcase, part 1
I did this in a semi-automated fashion. I used the F2 button in VS Code, which renames a variable. This should be less risky than doing it manually, but could still be prone to bugs 1) if it's a global variable, 2) if a variable of the same name already exists, etc.

If this deploy goes well and isn't buggy, I'll do the conversion to camelcase in more files in follow up patches.