speedy: improve message, refactor (#2244)
When Twinkle can't place a CSD tag on the mainspace page, it'll fall back to placing it on the talk page. But the message it leaves was a little confusing. I changed...

I was unable to tag TimedText:File:Foldergs.ogv.en.srt so please delete it.

... to ...

I was unable to tag TimedText:File:Foldergs.ogv.en.srt directly, so I have placed the speedy deletion tag on this page. I request deletion of the other page.

I also renamed a variable.